+++ Makefile
@@ -1,4 +1,4 @@
-VERSION := 3.24.0
+VERSION := 3.25.0
 PLUGINSLUG := woocart-defaults
 SRCPATH := $(shell pwd)/src
