+++ src/classes/Importers/ShippingZone.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace Niteo\WooCart\Defaults\Importers {
+
+	/**
+	 * Class ShippingZone
+	 *
+	 * @package Niteo\WooCart\Defaults\Importers
+	 */
+	class ShippingZone {
+
+
+		use FromArray;
+		use ToArray;
+
+		/**
+		 * @var int
+		 */
+		public $zone_id;
+		/**
+		 * @var string
+		 */
+		public $name;
+		/**
+		 * @var int
+		 */
+		public $order;
+		/**
+		 * @var array
+		 */
+		public $locations;
+		/**
+		 * @var array
+		 */
+		public $methods;
+	}
+
+}
